added ellipsis segment for routing to support several product formats
