+++ examples/commerce-essentials/src/app/(store)/products/[...productSegment]/product-display.tsx
